==> 1_dataCleaning.ipynb <==
import geopandas as gpd
import pandas as pd

# Importing TOPO1K as a GeoDataFrame

ibb = gpd.GeoDataFrame.from_file('data/ibb.geojson')

# Clear empty columns

empty_cols_ibb = [col for col in ibb.columns if ibb[col].isnull().all()]
ibb.drop(empty_cols_ibb, axis = 1, inplace = True)

# Add IDs to the features

Feature_ibb = []

for i in range(len(ibb)):
    Feature_ibb.append('ibb' + str(i+1))
    
ibb['fid'] = Feature_ibb

# Drop unnecessary columns

ibb.drop(['OBJECTID','GLOBALID','SHAPE_Leng','SHAPE_Area'],axis=1,inplace=True)

# building_num column has NaN values that cause datatype being float instead of an integer. So it needs to be fixed.

ibb['building_num'].fillna(0,inplace=True)
ibb['building_num'] = ibb['building_num'].astype('int64')

# Change special Turkish characters to Latin corresponding ones

tr2en = {'ğ':'g', 'ç':'c', 'ş':'s', 'ü':'u', 'ö':'o', 'ı':'i', 'Ğ':'G', 'Ç':'C', 'Ş':'S', 'Ü':'U', 'Ö':'O', 'İ':'I'}

for i in ibb.columns:
    try:
        ibb.replace({i:tr2en}, regex=True, inplace=True)
    except Exception:
        pass

#Import OSM as GeoDataFrame

osm = gpd.GeoDataFrame.from_file('data/osm.geojson')

# take only buildings of OSM data

osm = osm[osm["building"].notna()]

# take only buildings of OSM data

osm = osm[osm['building'] != 'transportation']

# clear empty columns

empty_cols_osm = [col for col in osm.columns if osm[col].isnull().all()]
osm.drop(empty_cols_osm, axis = 1, inplace = True)

# Change special Turkish characters to Latin correspondin ones

tr2en = {'ğ':'g', 'ç':'c', 'ş':'s', 'ü':'u', 'ö':'o', 'ı':'i', 'Ğ':'G', 'Ç':'C', 'Ş':'S', 'Ü':'U', 'Ö':'O', 'İ':'I'}

for i in osm.columns:
    try:
        osm.replace({i:tr2en}, regex=True, inplace=True)
    except Exception:
        pass

# Add IDs to the features

Feature_osm = []

for i in range(len(osm)):
    Feature_osm.append('osm' + str(i+1))
    
osm['fid'] = Feature_osm

# Arrange OSM ID column and delete '@' character from column name

osm['osmid'] = osm['id'].str.split('/', expand=True)[1]
osm.drop(['@id','id'],axis=1, inplace=True)

# Fix unwanted characters from column names

osm.columns = osm.columns.str.replace(':','_')
osm.columns = osm.columns.str.replace('@','')

# Fix format of timestamps of features

osm['timestamp'] = pd.to_datetime(osm['timestamp']).dt.strftime('%Y-%m-%d')

# Import TOPO25 data as GeoDataFrame

hgm = gpd.GeoDataFrame.from_file('data/hgm.geojson')

# Clear empty columns

empty_cols_hgm = [col for col in hgm.columns if hgm[col].isnull().all()]
hgm.drop(empty_cols_hgm, axis = 1, inplace = True)

# Add IDs to the features

Feature_hgm = []

for i in range(len(hgm)):
    Feature_hgm.append('hgm' + str(i+1))

hgm['fid'] = Feature_hgm

# Veri setinde yer alan Türkçe karakterleri latin karakterlerle değiştir.

tr2en = {'ğ':'g', 'ç':'c', 'ş':'s', 'ü':'u', 'ö':'o', 'ı':'i', 'Ğ':'G', 'Ç':'C', 'Ş':'S', 'Ü':'U', 'Ö':'O', 'İ':'I'}

for i in hgm.columns:
    try:
        hgm.replace({i:tr2en}, regex=True, inplace=True)
    except Exception:
        pass

# Export files as GeoJSON

ibb.to_file('data/clean_ibb.geojson', driver='GeoJSON')
osm.to_file('data/clean_osm.geojson', driver='GeoJSON')
hgm.to_file('data/clean_hgm.geojson', driver='GeoJSON')
