==> 2_matching.ipynb <==
import pandas as pd
import geopandas as gpd

# Import TOPO1K and OSM as GeoDataFrame

ibb = gpd.GeoDataFrame.from_file('data/clean_ibb.geojson')
osm = gpd.GeoDataFrame.from_file('data/clean_osm.geojson')

# Take desired columns into a new dataframe

ibb_int = ibb[['fid','geometry']].copy()
ibb_int['ibb_area'] = ibb_int['geometry'].area
osm_int = osm[['fid','geometry']].copy()
osm_int['osm_area'] = osm_int['geometry'].area

# Perform intersection analysis between TOPO1K and OSM

int_io = gpd.overlay(ibb_int, osm_int, how = 'intersection')

# Calculate areas and intesection ratio

int_io['int_area'] = int_io['geometry'].area
int_io['int_ibb'] = int_io['int_area']/int_io['ibb_area']
int_io['int_osm'] = int_io['int_area']/int_io['osm_area']
int_io['ratio'] = int_io['int_area']/(int_io[['ibb_area','osm_area']].min(axis=1))

# Identify features which has above 0.35 (35%) intersection ratio

int_io['match_ibb_osm'] = int_io['ratio'].apply(lambda x: 'Match' if x >= 0.35 else 'NoMatch')

# Take desired features to a new dataframe

match_ibb_osm = int_io[int_io['match_ibb_osm'] == 'Match'][['fid_1','fid_2']]

# reset index of new dataframe

match_ibb_osm.reset_index(drop=True)

# Export dataframe as CSV file

match_ibb_osm.to_csv('data/match_ibb_osm.csv')

# Import TOPO25K as GeoDataFrame

hgm = gpd.GeoDataFrame.from_file('data/clean_hgm.geojson')

# Split TOPO25K into two file. One contains buildings and another one contains settlement area 

hgm_yerlesim = hgm[hgm['feature_name'] == 'YERLESIM_YERI'] # Settlement area
hgm_bina = hgm[hgm['feature_name'] != 'YERLESIM_YERI'] # Building

# Take desired columns into a new dataframe

osm_int = osm[['fid','geometry']].copy()
osm_int['osm_area'] = osm_int['geometry'].area
hgm_int = hgm_bina[['fid','geometry']].copy()
hgm_int['hgm_area'] = hgm_int['geometry'].area

# Perform intersection analysis between OSM and TOPO25K buildings

int_oh = gpd.overlay(osm_int, hgm_int, how = 'intersection')

# Calculate areas and intesection ratio

int_oh['int_area'] = int_oh['geometry'].area
int_oh['int_osm'] = int_oh['int_area']/int_oh['osm_area']
int_oh['int_hgm'] = int_oh['int_area']/int_oh['hgm_area']
int_oh['ratio'] = int_oh['int_area']/(int_oh[['osm_area','hgm_area']].min(axis=1))

# Identify features which has above 0.35 (35%) intersection ratio

int_oh['match_osm_hgm'] = int_oh['ratio'].apply(lambda x: 'Match' if x >= 0.35 else 'NoMatch')

# Take desired features to a new dataframe

match_osm_hgm = int_oh[int_oh['match_osm_hgm'] == 'Match'][['fid_1','fid_2']]

# reset index of new dataframe

match_osm_hgm.reset_index(drop=True)

# Export dataframe as CSV file

match_osm_hgm.to_csv('data/match_osm_hgm.csv')

# Take desired columns into a new dataframe

hgm_int_yer = hgm_yerlesim[['fid','geometry']].copy()
hgm_int_yer['hgm_area'] = hgm_int_yer['geometry'].area

# Perform intersection analysis between OSM and TOPO25K settlement area

int_oh_yer = gpd.overlay(osm_int, hgm_int_yer, how = 'intersection')

# Calculate areas and intesection ratio

int_oh_yer['int_area'] = int_oh_yer['geometry'].area
int_oh_yer['int_osm'] = int_oh_yer['int_area']/int_oh_yer['osm_area']
int_oh_yer['int_hgm'] = int_oh_yer['int_area']/int_oh_yer['hgm_area']
int_oh_yer['ratio'] = int_oh_yer['int_area']/(int_oh_yer[['osm_area','hgm_area']].min(axis=1))

# Identify features which has above 0.35 (35%) intersection ratio

int_oh_yer['match_osm_hgm_yer'] = int_oh_yer['ratio'].apply(lambda x: 'Match' if x >= 0.35 else 'NoMatch')

# Take desired features to a new dataframe

match_osm_hgm_yer = int_oh_yer[int_oh_yer['match_osm_hgm_yer'] == 'Match'][['fid_1','fid_2']]

# reset index of new dataframe

match_osm_hgm_yer.reset_index(drop=True)

# Export dataframe as CSV file

match_osm_hgm_yer.to_csv('data/match_osm_hgm_yer.csv')
