==> 4_dataPropertiesToRDF.ipynb <==
import geopandas as gpd

# Import TOPO1K, OSM and TOPO25K as GeoDataFrames

ibb = gpd.GeoDataFrame.from_file('data/clean_ibb.geojson')
osm = gpd.GeoDataFrame.from_file('data/clean_osm.geojson')
hgm = gpd.GeoDataFrame.from_file('data/clean_hgm.geojson')

# take data columns

ibb_col = ibb.columns.tolist()
osm_col = osm.columns.tolist()
hgm_col = hgm.columns.tolist()

# create dataproperties from data columns as RDFs

cols = ibb_col + osm_col + hgm_col
cols = list(set(cols))

cols.remove('geometry')

dataProp = ['spo:has' + i[0].upper() + i[1:] + ' a owl:DatatypeProperty .' for i in cols]

with open('data/datatypeProperties.ttl', 'w') as output:
        for i in range(len(dataProp)):
            output.write('\n' + dataProp[i] + '\n')
