==> 5_featuresToRDF.ipynb <==
import pandas as pd
import geopandas as gpd
import rdfpandas as rpd

# Import TOPO1K as GeoDataFrame and prepare column names for RDFs 

ibb = gpd.GeoDataFrame.from_file('data/clean_ibb.geojson')

ibb['uri'] = 'http://ulusalharita.tr/id/topo#' + ibb['fid']
ibb['uri_geo'] = 'http://ulusalharita.tr/go/topo#' + ibb['fid']

ibb_geo = ibb[['uri_geo','geometry']].copy()
ibb_geo = ibb_geo.to_crs(epsg=4326)

ibb['uri_geo'] = '__' + ibb['uri_geo'] + '__'

ibb['rdfType'] = "owl:NamedIndividual"
ibb['class'] = "spo:BuildingAndFacilities"

ibb_geo['hasScale'] = "spo:largeScale"
ibb_geo['rdfType'] = "owl:NamedIndividual"
ibb_geo['class'] = "sf:Polygon"

ibb.drop(columns=['fid','geometry'], inplace=True)

# Build a list of URIs from column names for the creation of datatype properties

ibb_col = ibb.columns.tolist()
ls_ibb = ['http://ulusalharita.tr/ont/spo#' + 'has' + i[0].upper() + i[1:] for i in ibb_col]

# Create a dict for column rename

ibb_dict = dict(zip(ibb_col, ls_ibb))

# Remove unnecessary URIs from dict

ibb_dict.pop('uri')
ibb_dict.pop('uri_geo')
ibb_dict.pop('rdfType')
ibb_dict.pop('class')

# Rename column names according to the created dict

ibb.rename(columns=ibb_dict,inplace=True)

# Rename remainder column names

ibb.rename(columns={'uri_geo':'http://www.opengis.net/ont/geosparql#hasGeometry',
                    'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                    'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                    inplace=True)
ibb_geo.rename(columns={'geometry':'http://www.opengis.net/ont/geosparql#asWKT',
                        'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                        'hasScale':'http://ulusalharita.tr/ont/spo#hasScale',
                        'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                        inplace=True)

# Set index as feature URIs

ibb.set_index('uri',inplace=True)
ibb_geo.set_index('uri_geo',inplace=True)

# Import OSM as GeoDataFrame and prepare column names for RDFs 

osm = gpd.GeoDataFrame.from_file('data/clean_osm.geojson')

osm['uri'] = 'http://ulusalharita.tr/id/topo#' + osm['fid']
osm['uri_geo'] = 'http://ulusalharita.tr/go/topo#' + osm['fid']

osm_geo = osm[['uri_geo','geometry']].copy()
osm_geo = osm_geo.to_crs(epsg=4326)

osm['uri_geo'] = '__' + osm['uri_geo'] + '__'

osm['rdfType'] = "owl:NamedIndividual"
osm['class'] = "spo:Social_Culture"

osm_geo['hasScale'] = "spo:mediumScale"
osm_geo['rdfType'] = "owl:NamedIndividual"
osm_geo['class'] = "sf:Polygon"

osm.drop(columns=['fid','geometry'], inplace=True)

# Build a list of URIs from column names for the creation of datatype properties

osm_col = osm.columns.tolist()
ls_osm = ['http://ulusalharita.tr/ont/spo#' + 'has' + i[0].upper() + i[1:] for i in osm_col]

# Create a dict for column rename

osm_dict = dict(zip(osm_col, ls_osm))

# Remove unnecessary URIs from dict

osm_dict.pop('uri')
osm_dict.pop('uri_geo')
osm_dict.pop('rdfType')
osm_dict.pop('class')

# Rename column names according to the created dict

osm.rename(columns=osm_dict,inplace=True)

# Rename remainder column names

osm.rename(columns={'uri_geo':'http://www.opengis.net/ont/geosparql#hasGeometry',
                    'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                    'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                    inplace=True)

osm_geo.rename(columns={'geometry':'http://www.opengis.net/ont/geosparql#asWKT',
                        'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                        'hasScale':'http://ulusalharita.tr/ont/spo#hasScale',
                        'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                        inplace=True)

# Rename remainder column names

osm['http://ulusalharita.tr/ont/spo#hasTimestamp'] = osm['http://ulusalharita.tr/ont/spo#hasTimestamp'] + 'datatype2date'

# Set index as feature URIs

osm.set_index('uri',inplace=True)
osm_geo.set_index('uri_geo',inplace=True)

# Import TOPO25K as GeoDataFrame and prepare column names for RDFs 

hgm = gpd.GeoDataFrame.from_file('data/clean_hgm.geojson')

hgm['uri'] = 'http://ulusalharita.tr/id/topo#' + hgm['fid']
hgm['uri_geo'] = 'http://ulusalharita.tr/go/topo#' + hgm['fid']

hgm_geo = hgm[['uri_geo','geometry']].copy()
hgm_geo = hgm_geo.to_crs(epsg=4326)

hgm['uri_geo'] = '__' + hgm['uri_geo'] + '__'

hgm['rdfType'] = "owl:NamedIndividual"
hgm['class'] = "spo:Settlement"

hgm_geo['hasScale'] = "spo:smallScale"
hgm_geo['rdfType'] = "owl:NamedIndividual"
hgm_geo['class'] = "sf:Polygon"

hgm.drop(columns=['fid','geometry','shape_leng','shape_area'], inplace=True)

# Build a list of URIs from column names for the creation of datatype properties

hgm_col = hgm.columns.tolist()
ls_hgm = ['http://ulusalharita.tr/ont/spo#' + 'has' + i[0].upper() + i[1:] for i in hgm_col]

# Create a dict for column rename

hgm_dict = dict(zip(hgm_col, ls_hgm))

# Remove unnecessary URIs from dict

hgm_dict.pop('uri')
hgm_dict.pop('uri_geo')
hgm_dict.pop('rdfType')
hgm_dict.pop('class')

# Rename column names according to the created dict

hgm.rename(columns=hgm_dict,inplace=True)

# Rename remainder column names

hgm.rename(columns={'uri_geo':'http://www.opengis.net/ont/geosparql#hasGeometry',
                    'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                    'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                    inplace=True)

hgm_geo.rename(columns={'geometry':'http://www.opengis.net/ont/geosparql#asWKT',
                        'rdfType':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
                        'hasScale':'http://ulusalharita.tr/ont/spo#hasScale',
                        'class':'http://www.w3.org/1999/02/22-rdf-syntax-ns#type'},
                        inplace=True)

# Set index as feature URIs

hgm.set_index('uri',inplace=True)
hgm_geo.set_index('uri_geo',inplace=True)

# Correct data types of newly created dataframes

ibb = ibb.apply(pd.to_numeric, errors='ignore')
ibb_geo = ibb_geo.apply(pd.to_numeric, errors='ignore')
osm = osm.apply(pd.to_numeric, errors='ignore')
osm_geo = osm_geo.apply(pd.to_numeric, errors='ignore')
hgm = hgm.apply(pd.to_numeric, errors='ignore')
hgm_geo = hgm_geo.apply(pd.to_numeric, errors='ignore')

# Replace all null values to 'None'

import numpy as np

ibb     = ibb.replace({np.NaN: None})
ibb_geo = ibb_geo.replace({np.NaN: None})
osm     = osm.replace({np.NaN: None})
osm_geo = osm_geo.replace({np.NaN: None})
hgm     = hgm.replace({np.NaN: None})
hgm_geo = hgm_geo.replace({np.NaN: None})

# Convert all created dataframes to RDF Graph

graph = rpd.to_graph(ibb) + rpd.to_graph(ibb_geo) + rpd.to_graph(osm) +
        rpd.to_graph(osm_geo) + rpd.to_graph(hgm) + rpd.to_graph(hgm_geo)

# Add necessary prefixes

graph.bind('geo', 'http://www.opengis.net/ont/geosparql#')
graph.bind('spo', 'http://ulusalharita.tr/ont/spo#')
graph.bind('idt', 'http://ulusalharita.tr/id/topo#')
graph.bind('got', 'http://ulusalharita.tr/go/topo#')

# Remove all triples that has None value as predicate

from rdflib import Literal

graph.remove((None, None, Literal("None")))

# Serialize graph as Turtle format

graph.serialize('data/individuals.ttl',format='turtle')

# Open created Turtle file and make last modifications for a consistent file and write all data to a file as individuals

with open('data/individuals.ttl','r') as ind:
    inddata = ind.read()
    inddata = inddata.replace('"owl:NamedIndividual"','owl:NamedIndividual')
    inddata = inddata.replace('"spo:smallScale"','spo:smallScale')
    inddata = inddata.replace('"spo:mediumScale"','spo:mediumScale')
    inddata = inddata.replace('"spo:largeScale"','spo:largeScale')
    inddata = inddata.replace('"sf:Polygon"','sf:Polygon')
    inddata = inddata.replace('"spo:Social_Culture"','spo:Social_Culture')
    inddata = inddata.replace('"spo:Settlement"','spo:Settlement')
    inddata = inddata.replace('"spo:BuildingAndFacilities"','spo:BuildingAndFacilities')
    inddata = inddata.replace('))" ','))"^^geo:wktLiteral ')
    inddata = inddata.replace('datatype2date"','"^^xsd:date')
    inddata = inddata.replace('"__','<')
    inddata = inddata.replace('__"','>')
    
with open('data/individuals.ttl','w') as file:
    file.write(inddata)
    
with open('data/individuals.ttl','r') as fin:
    data = fin.read().splitlines(True)
    
with open('data/individuals.ttl','w') as fout:
    fout.writelines(data[5:])
