==> 7_mergeAll.ipynb <==
# Merge all Turtle files into the Geospatial Ontology

with open('data/SpatialOnt_HEAD.ttl','r') as ont:
    ontology = ont.read()
    
with open('data/datatypeProperties.ttl','r') as data:
    dataProperties = data.read()

with open('data/individuals.ttl','r') as ind:
    indviduals = ind.read()

with open('data/relations.ttl','r') as rel:
    relations = rel.read()
    
with open('data/links.ttl','r') as lin:
    links = lin.read()
    
with open('data/SpatialOnt.ttl','w') as file:
    file.write(ontology + dataProperties + indviduals + relations + links)

# Check file if there is an error

from rdflib import Graph

g = Graph()
g.parse("data/SpatialOnt.ttl",  format='turtle')

g.serialize('data/SpatialOnt.ttl', format='turtle')
